+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Possibility to view channel info on channel options
 - Date separators in the message list
 - ChatUserNamer to customize user name on typing indicator
+- minimumSwipeGestureDistance to control swipe sensitivity
 
 ### 🐞 Fixed
 - Bug about link attachments not opening when the URL was missing the scheme